refactored lightGBM to use SparkML base classes
